remove cdbwiki
we've not been using it for years, and strictly speaking I would have
to split it from mwlib anyway when doing the debian packaging for
mwlib according to wikimedia foundation's brain-damaged guidelines.

if someone still needs this and wants to maintain a mwlib.cdb package,
please speak up on the mailing list.